* register dark images
git-svn-id: http://svn.code.sf.net/p/rts-2/code/branches/rts-2/REL_0_5_0_exp@1596 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/sql/create_tables.sql
@@ -158,8 +158,8 @@ DROP TABLE darks;
 CREATE TABLE darks (
 	dark_name	varchar(250) NOT NULL,
 	dark_date	timestamp,
-	dark_exposure	integer,
-	dark_temperature integer,
+	dark_exposure	float,
+	dark_temperature float,
 	epoch_id	integer NOT NULL REFERENCES epoch(epoch_id),
 	camera_name	varchar(8) REFERENCES cameras(camera_name)
 );
